done fix delete consultantbank

+++ Monhealth.Application/Features/ConsultantBank/Commands/DeleteConsultantBank/DeleteConsultantBankCommandHandler.cs
@@ -13,6 +13,12 @@ namespace Monhealth.Application.Features.ConsultantBank.Commands.DeleteConsultan
             {
                 return false;
             }
+            var consultantBanks = await consultantBankRepository.GetConsultantBankByConsultantIdAsync(consultantBank.ConsultantId);
+
+            if (consultantBanks.Count() <= 1)
+            {
+                throw new Exception("Không thể xóa, bạn phải liên kết ít nhất 1 ngân hàng");
+            }
             consultantBankRepository.Remove(consultantBank);
             await consultantBankRepository.SaveChangeAsync(cancellationToken);
             return true;
